+++ src/main/java/gift/presentation/PageController.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.presentation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
